+++ R/bruceR-stats_3_manova.R
@@ -738,8 +738,8 @@ MANOVA = function(data, subID=NULL, dv=NULL,
 #'   EMMEANS("A", by="B") %>%
 #'   EMMEANS("B", by="A")
 #' ## How to create an interaction plot using `emmeans::emmip()`?
-#' ## See help page for its usage: ?emmeans::emmip()
-#' m = MANOVA(between.2, dv="SCORE", between=c("A", "B"))
+#' ## See help page: ?emmeans::emmip()
+#' \donttest{m = MANOVA(between.2, dv="SCORE", between=c("A", "B"))
 #' emmip(m, ~ A | B, CIs=TRUE)
 #' emmip(m, ~ B | A, CIs=TRUE)
 #' emmip(m, B ~ A, CIs=TRUE)
@@ -752,7 +752,7 @@ MANOVA = function(data, subID=NULL, dv=NULL,
 #'   EMMEANS("A", by=c("B", "C"))
 #' ## Just to name a few...
 #' ## You may test other combinations...
-#'
+#' }
 #'
 #' #### Within-Subjects Design ####
 #'
@@ -761,7 +761,7 @@ MANOVA = function(data, subID=NULL, dv=NULL,
 #'        within="A") %>%
 #'   EMMEANS("A")
 #'
-#' within.2
+#' \donttest{within.2
 #' MANOVA(within.2, dvs="A1B1:A2B3", dvs.pattern="A(.)B(.)",
 #'        within=c("A", "B")) %>%
 #'   EMMEANS("A", by="B") %>%
@@ -771,7 +771,7 @@ MANOVA = function(data, subID=NULL, dv=NULL,
 #' # the linear dependence of A2B2 and A2B3.
 #' # See: Corr(within.2[c("A2B2", "A2B3")])
 #'
-#' \donttest{within.3
+#' within.3
 #' MANOVA(within.3, dvs="A1B1C1:A2B2C2", dvs.pattern="A(.)B(.)C(.)",
 #'        within=c("A", "B", "C")) %>%
 #'   EMMEANS("A", by="B") %>%