+++ R/bruceR-stats_5_advance.R
@@ -442,9 +442,7 @@ boot_ci = function(boot,
 #' \emph{Journal of Personality and Social Psychology, 115}(6), 929-943.
 #'
 #' @examples
-#' \dontrun{
-#'
-#' #### NOTE ####
+#' \donttest{#### NOTE ####
 #' ## In the following examples, I set nsim=100 to save time.
 #' ## In formal analyses, nsim=1000 (or larger) is suggested!
 #'
@@ -538,7 +536,6 @@ boot_ci = function(boot,
 #' ## For more examples and details, see the "note" subfolder at:
 #' ## https://github.com/psychbruce/bruceR/tree/main/note
 #' }
-#'
 #' @export
 PROCESS = function(data,
                    y="",